% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starnote.R
\name{starnote}
\alias{starnote}
\title{Create tabularx tables using stargazer}
\usage{
starnote(
  dta = "",
  save_loc = "",
  tablenote = "",
  colnumber = 6,
  width_col1 = 0.15
)
}
\arguments{
\item{dta}{save_loc tablenote colnumber}
}
\value{
.tex file
}
\description{
Transforms stargazer output into textwidth tabularx and adds a wrapping table note
}
\examples{
TBD

}
